gralloc: Do not map framebuffer memory unless needed
The SurfaceFlinger opens the framebuffer module even when it
isn't necessary. It does this for some legacy hals.
Due to this we unnecessarily mmap the fb memory in gralloc.
Do not map this unless the property to use the framebuffer
memory is set.

We do further operations in fb_device_open based on the result of
mapFrameBuffer. However, if mapFrameBuffer is never done, we
could end up with a random crash when using the info values in
fb_device_open. Defaulting to an error value avoids this.

Change-Id: Ib4ebfdf9a63af0dabb53170342181bac0360baeb

Change-Id: I1e51f814611a3f2eb73f36d75e14d8725ca98474